+++ tests/testthat/test_pipelines/test_pipeline_days_bad.R
@@ -1,0 +1,5 @@
+#' @maestroFrequency hourly
+#' @maestroDays Mon Wed
+some_days <- function() {
+  invisible()
+}