+++ tests/testthat/test_pipelines/test_pipeline_days_bad2.R
@@ -1,0 +1,5 @@
+#' @maestroFrequency hourly
+#' @maestroDays Mon 4
+some_days <- function() {
+  invisible()
+}