+++ tests/testthat/test_pipelines/test_pipeline_days_good.R
@@ -1,0 +1,5 @@
+#' @maestroFrequency hourly
+#' @maestroDays 1 14
+some_days <- function() {
+  invisible()
+}