+++ tests/testthat/test_pipelines/test_pipeline_hours_bad.R
@@ -1,0 +1,5 @@
+#' @maestroFrequency hourly
+#' @maestroHours a
+one_hour <- function() {
+  invisible()
+}