+++ tests/testthat/test_pipelines_parse_all_good/specifiers.R
@@ -1,0 +1,43 @@
+#' @maestroFrequency daily
+#' @maestroDays 1 8 24 28
+specific_days <- function() {
+
+}
+
+#' @maestroFrequency daily
+#' @maestroDays Tue Fri
+specific_days2 <- function() {
+
+}
+
+#' @maestroFrequency hourly
+#' @maestroHours 0 4 8 12 16 20
+specific_hours <- function() {
+
+}
+
+#' @maestroFrequency biweekly
+#' @maestroMonths 1 5 10
+specific_months1 <- function() {
+
+}
+
+#' @maestroFrequency weekly
+#' @maestroMonths 1 5 10
+specific_months2 <- function() {
+
+}
+
+#' @maestroFrequency monthly
+#' @maestroMonths 1 5 10
+specific_months3 <- function() {
+
+}
+
+#' @maestroFrequency hourly
+#' @maestroHours 1 12
+#' @maestroDays 1
+#' @maestroMonths 1 5 10
+specific_multi <- function() {
+
+}